requirements.txt
Django==5.1.4
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
social-auth-app-django==5.4.0
Pillow==10.1.0
python-dotenv==1.0.0
django-cleanup==8.0.0
django-widget-tweaks==1.5.0
django-allauth==0.59.0
django-environ==0.11.2
django-storages==1.14.2
django-debug-toolbar==4.2.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9